Refactored where some files reside. Created new user_interaction subdirectory
