+++ src/driver.cpp
@@ -18,8 +18,8 @@
 #include "device/RangeSensorContainer.h"
 #include "device/sensors_encoders.h"
 #include "legacy_motion/motion.h"
-#include "FreakOut.h"
-#include "Menu.h"
+#include "user_interaction/FreakOut.h"
+#include "user_interaction/Menu.h"
 #include "conf.h"
 #include "data.h"
 #include "parser.h"
