+++ src/legacy_motion/motion.cpp
@@ -3,12 +3,12 @@
 #include "../device/Orientation.h"
 #include "../device/RangeSensor.h"
 #include "../device/RangeSensorContainer.h"
-#include "../device/UserInterface.h"
+#include "../user_interaction/UserInterface.h"
 #include "../device/sensors_encoders.h"
 #include "../conf.h"
-#include "../FreakOut.h"
-#include "../Logger.h"
-#include "../Menu.h"
+#include "../user_interaction/FreakOut.h"
+#include "../user_interaction/Logger.h"
+#include "../user_interaction/Menu.h"
 #include "MotionCalc.h"
 #include "PIDController.h"
 #include "SweptTurnProfile.h"
