+++ src/main.cpp
@@ -10,14 +10,14 @@
 #include "device/Motor.h"
 #include "device/Orientation.h"
 #include "device/PersistantStorage.h"
-#include "device/PlayMelodies.h"
+#include "user_interaction/PlayMelodies.h"
 #include "device/RangeSensorContainer.h"
-#include "device/UserInterface.h"
+#include "user_interaction/UserInterface.h"
 #include "device/sensors_encoders.h"
 #include "legacy_motion/motion.h"
 #include "Log.h"
-#include "Logger.h"
-#include "Menu.h"
+#include "user_interaction/Logger.h"
+#include "user_interaction/Menu.h"
 #include "Navigator.h"
 #include "conf.h"
 #include "data.h"
