+++ src/user_interaction/FreakOut.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
-#include "device/Motor.h"
-#include "device/PlayMelodies.h"
-#include "device/UserInterface.h"
+#include "../device/Motor.h"
+#include "PlayMelodies.h"
+#include "UserInterface.h"
 #include "Logger.h"
 #include "Menu.h"
 #include "FreakOut.h"
