+++ src/user_interaction/Menu.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
-#include "device/PersistantStorage.h"
-#include "device/UserInterface.h"
-#include "driver.h"
+#include "../device/PersistantStorage.h"
+#include "UserInterface.h"
+#include "../driver.h"
 #include "Menu.h"
 
 static void doNothing() {};
